fixed distance bug and refactored code
